Crate osdk_heap_allocator

Structs

CpuLocalBox

A wrapper over DynamicCpuLocal<T> to deallocate CPU-local objects on drop automatically.

HeapAllocator

The global heap allocator provided by OSDK.

Functions

alloc_cpu_local

Allocates a dynamically-allocated CPU-local object of type T and initializes it with init_values.

type_from_layout

Get the type of the slot from the layout.