osdk_heap_allocator

Function alloc_cpu_local

pub fn alloc_cpu_local<T>(
    init_values: impl FnMut(CpuId) -> T,
) -> Result<CpuLocalBox<T>>

Allocates a dynamically-allocated CPU-local object of type T and initializes it with init_values.

Currently, the size of T must be no larger than 32 bytes.