osdk_heap_allocator

Struct HeapAllocator

pub struct HeapAllocator;

The global heap allocator provided by OSDK.

It is a singleton that provides heap allocation for the kernel. If multiple instances of this struct are created, all the member functions will eventually access the same allocator.